PCFG - script to generate seed data
This script takes preprocessed training data
and extracts review text - useful and not useful reviews
seperately based on arguments supplied. The output generated
by this script can be fed to parser for producing parses
of the review text. These parses are then used to train
seperate parsers for useful and not useful reviews.

Signed-off-by: Vidhoon Viswanathan <vidhoon@cs.utexas.edu>